Decouple media button starter and receiver
