+++ app/src/main/java/de/danoeh/antennapod/adapter/actionbutton/PauseActionButton.java
@@ -5,10 +5,10 @@ import android.view.KeyEvent;
 import androidx.annotation.DrawableRes;
 import androidx.annotation.StringRes;
 import de.danoeh.antennapod.R;
-import de.danoeh.antennapod.core.receiver.MediaButtonReceiver;
 import de.danoeh.antennapod.core.util.PlaybackStatus;
 import de.danoeh.antennapod.model.feed.FeedItem;
 import de.danoeh.antennapod.model.feed.FeedMedia;
+import de.danoeh.antennapod.ui.appstartintent.MediaButtonStarter;
 
 public class PauseActionButton extends ItemActionButton {
 
@@ -36,7 +36,7 @@ public class PauseActionButton extends ItemActionButton {
         }
 
         if (PlaybackStatus.isCurrentlyPlaying(media)) {
-            context.sendBroadcast(MediaButtonReceiver.createIntent(context, KeyEvent.KEYCODE_MEDIA_PAUSE));
+            context.sendBroadcast(MediaButtonStarter.createIntent(context, KeyEvent.KEYCODE_MEDIA_PAUSE));
         }
     }
 }
