+++ core/src/main/java/de/danoeh/antennapod/core/receiver/MediaButtonReceiver.java
@@ -1,10 +1,8 @@
 package de.danoeh.antennapod.core.receiver;
 
-import android.app.PendingIntent;
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
-import android.os.Build;
 import androidx.core.content.ContextCompat;
 import android.util.Log;
 import android.view.KeyEvent;
@@ -22,7 +20,6 @@ public class MediaButtonReceiver extends BroadcastReceiver {
     public static final String EXTRA_SOURCE = "de.danoeh.antennapod.core.service.extra.MediaButtonReceiver.SOURCE";
     public static final String EXTRA_HARDWAREBUTTON
             = "de.danoeh.antennapod.core.service.extra.MediaButtonReceiver.HARDWAREBUTTON";
-    public static final String NOTIFY_BUTTON_RECEIVER = "de.danoeh.antennapod.NOTIFY_BUTTON_RECEIVER";
     public static final String PLAYBACK_SERVICE_INTENT = "de.danoeh.antennapod.intents.PLAYBACK_SERVICE";
 
     @Override
@@ -47,16 +44,4 @@ public class MediaButtonReceiver extends BroadcastReceiver {
         }
     }
 
-    public static Intent createIntent(Context context, int eventCode) {
-        KeyEvent event = new KeyEvent(KeyEvent.ACTION_DOWN, eventCode);
-        Intent startingIntent = new Intent(context, MediaButtonReceiver.class);
-        startingIntent.setAction(MediaButtonReceiver.NOTIFY_BUTTON_RECEIVER);
-        startingIntent.putExtra(Intent.EXTRA_KEY_EVENT, event);
-        return startingIntent;
-    }
-
-    public static PendingIntent createPendingIntent(Context context, int eventCode) {
-        return PendingIntent.getBroadcast(context, eventCode, createIntent(context, eventCode),
-                (Build.VERSION.SDK_INT >= 23 ? PendingIntent.FLAG_IMMUTABLE : 0));
-    }
 }
